+++ AWSElasticLoadBalancing/AWSElasticLoadBalancingService.m
@@ -26,7 +26,7 @@
 #import "AWSElasticLoadBalancingResources.h"
 
 static NSString *const AWSInfoElasticLoadBalancing = @"ElasticLoadBalancing";
-static NSString *const AWSElasticLoadBalancingSDKVersion = @"2.6.15";
+static NSString *const AWSElasticLoadBalancingSDKVersion = @"2.6.16";
 
 
 @interface AWSElasticLoadBalancingResponseSerializer : AWSXMLResponseSerializer