+++ AWSKinesis/AWSFirehoseService.m
@@ -26,7 +26,7 @@
 #import "AWSFirehoseResources.h"
 
 static NSString *const AWSInfoFirehose = @"Firehose";
-static NSString *const AWSFirehoseSDKVersion = @"2.6.15";
+static NSString *const AWSFirehoseSDKVersion = @"2.6.16";
 
 
 @interface AWSFirehoseResponseSerializer : AWSJSONResponseSerializer