+++ AWSPolly/AWSPollyService.m
@@ -26,7 +26,7 @@
 #import "AWSPollyResources.h"
 
 static NSString *const AWSInfoPolly = @"Polly";
-static NSString *const AWSPollySDKVersion = @"2.6.15";
+static NSString *const AWSPollySDKVersion = @"2.6.16";
 
 
 @interface AWSPollyResponseSerializer : AWSJSONResponseSerializer