+++ AWSSQS/AWSSQSService.m
@@ -26,7 +26,7 @@
 #import "AWSSQSResources.h"
 
 static NSString *const AWSInfoSQS = @"SQS";
-static NSString *const AWSSQSSDKVersion = @"2.6.15";
+static NSString *const AWSSQSSDKVersion = @"2.6.16";
 
 
 @interface AWSSQSResponseSerializer : AWSXMLResponseSerializer