+++ Scripts/GenerateAppleDocs.sh
@@ -9,7 +9,7 @@ function cleanup
 }
 
 
-VERSION="2.6.15"
+VERSION="2.6.16"
 if [ -n $1 ] && [ "$1" == "clean" ];
 then
 	cleanup